migrate cf-component-tabs tests to Jest
